@@ -23,9 +23,9 @@ public class SecurityConfiguration {
             .csrf()
             .disable()
             .authorizeRequests()
-            .antMatchers("/user/expense/**")
+            .requestMatchers("/user/expense/**")
             .hasRole("USER")
-            .antMatchers("/user/register")
+            .requestMatchers("/user/register")
             .permitAll()
             .and()
             .httpBasic(Customizer.withDefaults())
